+++ secreto/src/main/java/com/emelmujiro/secreto/notification/dto/request/GetAllNotificationsRequestDto.java
@@ -1,7 +1,6 @@
 package com.emelmujiro.secreto.notification.dto.request;
 
 import com.emelmujiro.secreto.auth.annotation.LoginUser;
-import com.emelmujiro.secreto.mission.dto.request.NotificationSearchPeriod;
 import lombok.*;
 import org.springframework.data.domain.Pageable;
