+++ secreto/src/main/java/com/emelmujiro/secreto/notification/dto/request/GetRoomNotificationsRequestDto.java
@@ -2,10 +2,8 @@ package com.emelmujiro.secreto.notification.dto.request;
 
 import com.emelmujiro.secreto.auth.annotation.LoginUser;
 import com.emelmujiro.secreto.global.annotation.InjectPathVariable;
-import com.emelmujiro.secreto.mission.dto.request.NotificationSearchPeriod;
 import lombok.*;
 import org.springframework.data.domain.Pageable;
-import org.springframework.web.bind.annotation.PathVariable;
 
 @Setter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
